add some comments, move update_recent_quota() to quota.rs
